.
added ur wall stuff, not working yet tho, changed float pos to int pos

+++ GAME/Coin.pde
@@ -1,7 +1,7 @@
 public class Coin extends GamePiece {
   private color coinColor = #FFFF00;
   
-  public Coin(float xpos,float ypos) {
+  public Coin(int xpos,int ypos) {
     super(xpos, ypos);
   }
